requirements.txt
# Backend
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
chromadb==0.5.5
pypdf==4.3.1
sentence-transformers==3.0.1
langchain==0.2.12
langchain-community==0.2.11
langchain-text-splitters==0.2.2
# If you prefer OpenAI embeddings/LLM (uncomment in code):
openai==1.40.3

# Web retrieval (PubMed + parsing)
biopython>=1.85
beautifulsoup4==4.12.3
requests==2.32.3

# Production server
gunicorn==21.2.0
